requirements.txt
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
praw==7.7.1
pmaw==3.0.0
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
textblob==0.17.1
emoji==2.8.0
contractions==0.1.73
wordcloud==1.9.3
matplotlib==3.8.2
seaborn==0.13.0
vaderSentiment==3.3.2
plotly==5.18.0
